fixup! Add code towards having linear bivariate regression
`length something` returns an `Int`, `fromIntegral` converts it to a
`Double`.

+++ Stat.hs
@@ -36,7 +36,7 @@ pointCost mf (x, y) = diff^2
 
 {- | Calculate cost of some model over some data. -}
 cost :: Model -> Data -> Double
-cost m d = sum (map pointCost' d) / (2 * length d)
+cost m d = sum (map pointCost' d) / (2 * fromIntegral (length d))
   where
     mf = toModelF m
     pointCost' = pointCost mf
